+++ CHANGELOG.md
@@ -10,7 +10,9 @@ Mark  items as `Added`, `Changed`, `Fixed`, `Modified`, `Removed`, `Untested Fea
 **forms-flow-api**
 
 
-* Added a new `/form-flow-builder` endpoint to handle form, workflow, authorization, and settings creation.
+* Added new endpoints to manage form flow creation and updates:
+   * `/form-flow-builder` - Create form design, workflow, authorizations, and mapper details.
+   * `/form-flow-builder/<mapper-id>` - Update existing configurations.
 
 
 ## 7.3.0 - Date To be updated